% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_SchemaRedirect.r
\name{hash_data_structure.SchemaRedirect_v8}
\alias{hash_data_structure.SchemaRedirect_v8}
\title{Hashing data structure (SchemaRedirect_v8)}
\usage{
\method{hash_data_structure}{SchemaRedirect_v8}(x, col, ...)
}
\arguments{
\item{x}{A SchemaRedirect_v8 object}

\item{col}{The column to be hashed}

\item{...}{Unused}
}
\description{
An implementation of spltidy::hash_data_structure for SchemaRedirect_v8
}
